flist: add flist_first_entry()

+++ gclient.c
@@ -694,7 +694,7 @@ static void gfio_client_job_start(struct fio_client *client, struct fio_net_cmd 
 
 static void gfio_client_iolog(struct fio_client *client, struct cmd_iolog_pdu *pdu)
 {
-	printf("got iolog: name=%s, type=%u, entries=%u\n", pdu->name, pdu->log_type, pdu->nr_samples);
+	printf("got iolog: name=%s, type=%u, entries=%lu\n", pdu->name, pdu->log_type, (unsigned long) pdu->nr_samples);
 	free(pdu);
 }
